moved interactive logic to utils file
